+++ smartcontract/cli/src/globalconfig/authority/get.rs
@@ -61,6 +61,7 @@ mod tests {
             device_tunnel_block: "10.1.0.0/24".parse().unwrap(),
             user_tunnel_block: "10.5.0.0/24".parse().unwrap(),
             multicastgroup_block: "224.2.0.0/4".parse().unwrap(),
+            next_bgp_community: 10000,
         };
 
         client
@@ -74,7 +75,7 @@ mod tests {
         assert!(res.is_ok());
         let output_str = String::from_utf8(output).unwrap();
         assert_eq!(
-            output_str, " local asn | remote asn | device tunnel block | user tunnel block | multicast group block \n 1234      | 5678       | 10.1.0.0/24         | 10.5.0.0/24       | 224.2.0.0/4           \n"
+            output_str, " local asn | remote asn | device tunnel block | user tunnel block | multicast group block | next bgp community \n 1234      | 5678       | 10.1.0.0/24         | 10.5.0.0/24       | 224.2.0.0/4           | 10000              \n"
         );
     }
 }