+++ CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 
 ### Changes
 
+- Onchain programs
+    - serviceability: add auto-assignment and validation for exchange.bgp_community
+- CLI
+    - Removed `--bgp-community` option from `doublezero exchange create` since these values are now assigned automatically
+    - Add `--next-bgp-community` option to `doublezero global-config set` so authorized users can control which bgp_community will be assigned next
+
 ## [v0.6.8](https://github.com/malbeclabs/doublezero/compare/client/v0.6.6...client/v0.6.8) – 2025-10-17
 
 ### Breaking